New Conf for Auth
Added a new option IMongoCmdOptions and on its implementation and builder to have a new option that the user can set so the authentication of MongoDB will work. Changed Mongod to handle this feature. I didn't find any specific test for the options and I thought it would be harder to understand the test structure of the project so I didn't include a test with this PR unfurtunately (by the way it took a lot of time to run the test in my machine and some fail and I created this branch from master) .